Create additional divs and end game and play again
Create more divs to replace the console logs and structure the information better on the browser. Create a function to define the end of the game and a way for the user to start the game over.

Need to figure out why the browser doesn't populate anything after restarting the game.